Initial checkin of rcpt_ext_script
This plugin is supposed to call an external script which purpose is
to check the recipient. If the external script returns 0, the given
recipient/mailbox is valid. In case != 0 is returned,
the given recipient has not been found.
This is an initial version derived from qmail_ldap_deliverable script
which doesn't allow yet to specify the script name as parameter.