Set script via parameter
The script to be called for recipient verification can be set in
plugins config file. The parameter is called 'script' and the value
must contain the full path to the script. Example:

rcpt_ext_script script /var/qmail/bin/rcptcheck.sh

The script gets called with the recipient address as input parameter
and has to return 0 in case th recipient is value, all other return
values mean the recipient is invalid.